OBBs, geo tests, eigensolver, and unit tests.
Designed, implemented and tested an oriented bounding box interface in
primal/src/geometry. Also designed, implemented and tested intersection tests
for OBB-OBBs in 2 and 3 dimensions in primal/src/operators. Finally: designed,
implemented and tested a constructor from a collection of points. This last
constructor uses the principal directions from the collection of points to
determine the orientation of the OBB, and thus required an eigen solver.
The solver was implemented in axom_utils/src/numerics. All of this was
unit tested in the appropriate directories. All unit tests are currently
passing.

Committer: Austin James Stromme <stromme1@cab689.llnl.gov>

On branch feature/stromme1/OBB
Your branch is up-to-date with 'origin/feature/stromme1/OBB'.

Changes to be committed:
 modified:   src/components/axom_utils/src/numerics/eigen_solve.hpp
 modified:   src/components/axom_utils/tests/numerics_eigen_solve.cpp
 modified:   src/components/primal/src/CMakeLists.txt
 modified:   src/components/primal/src/geometry/OrientedBoundingBox.hpp
 modified:   src/components/primal/src/operators/closest_point.hpp
 new file:   src/components/primal/src/operators/compute_bounding_box.hpp
 modified:   src/components/primal/src/operators/intersect.hpp
 modified:   src/components/primal/tests/CMakeLists.txt
 new file:   src/components/primal/tests/primal_compute_bounding_box.cpp
 modified:   src/components/primal/tests/primal_intersect.cpp
 modified:   src/components/primal/tests/primal_orientedboundingbox.cpp

Untracked files:
 uberenv_libs/