DOC: clarify guidelines for doxygen documentation.
The intent for Policy 7.12 is for when editing existing files
to ensure that doxygen style is consistent within the file.
However, it is misinterpeted as "if I add a new file, I can
change to whichever Doxygen style I like", which is problematic
because it leads to inconsistent style within a component, which
is bad.

This commit clarifies that 7.12 applies to existing files and
explicitly adds Policy 7.13 to require a consistent doxygen
style within a component.

This resolves ATK-1071.