Adds an ImplicitGrid spatial index to quest. ATK-962.
An ImplicitGrid covers an axis-aligned cube of space (in 1D, 2D, 3D, or higher)
and tracks the occupancy of its rectilinear grid cells with respect to an index set of elements
via 1D projections along bins associated with the coordinate axes.

Each bin stores a bitset (one bit per element) to track the occupancy of the 1D projection of an
element's bounding box. The storage of an ImplicitGrid is O( N * (b_x [+ b_y [+ b_z] ] ) ),
where N is the number of elementx, and b_x, b_y and b_z are the number of bins
along the x-, y- and z- dimensions.