ENH: Add Axom Version info in generated config.hpp
Adds Axom Version information to the generated config
header. This commit leverages new git macros that are
added in BLT to autodetect the version information from
Git tag when we are building from a Git repo and Git is
available.

Also added missing and corrected copyright info in *.cmake files.

This commit resolves ATK-941.