Added a facility to read in files in the source tree.
From Cyrus Harrison's Conduit project, I added a facility to store the
absolute path to the source tree in a header file.  This is used by the
mesh tester tests to read in various test definition files and test meshes.
Several of these definition files and their mesh files are included.