ENH: construct an FE object from a Matrix
Adds the ability to construct an FE object from a matrix
consisting of the element coordinates, where, the number
of rows corresponds to the dimension of the element and
the number of columns to the number of nodes in the element.
Moreover, the caller can specify whether the coordinate will
be shallow-copied or deep-copied to the FE object internals.

This decouples the FE object from the mint::Mesh and allows
the code to be more readily used by other codes that may
use a different mesh data model.

This commit resolves ATK-1056.