ENH: support non-watertight input geometry
This commit adds support to Quest's signed distance query for
input consisting of a surface mesh that is not closed, i.e.,
a compact manifold with boundary. The initial implementation
assumed the input was a closed surface. However, for certain
applications, this assumption does not hold true, in which
case, the code evaluates the magnitude of the distance correctly,
but, computes the sign incorrectly.

The API for the signed distance query is also updated to allow
the caller to toggle on whether the input consists of a closed
surface or not.

Last, a test for non-watertight input geometry is added to the
signed_distance_interface unit tests.

This commit resolves ATK-1100.