Modified implementation of PointInCell's mfem mesh wrapper.
PointInCell requires a nodal grid function with a positive
(Bernstein) basis to compute a bounding box that is guaranteed to
cover the element. Rather than retaining a nodal grid function
and recomputing the bounding boxes at each query, we are now
caching the element bounding boxes. This makes queries faster
and reduces the storage overhead when the original mesh's
grid function is not positive (the anticipated general case).

Also reduced the geometric refinement order for initializing
mfem's TransformBack function on each candidate element from 2 to 1.
This uses a factor of two fewer isoparameter-to-space transformations
(per dimension) without introducing errors when applied to linear
and quadratic meshes in 2D and 3D.

Also adds a PointInCell test for randomly jittered quadratic hexahedral meshes.