Properly handle initializing the grid resolution in quest's ImplicitGrid.
When a user passes in a NULL pointer for the grid resolution, we use a
heuristic to set the grid resolution as the N^th root of the number of
elements in the index set (where N is the dimension).  We also ensure
that each dimension's resolution is at least one.