Modifies calls to use mfem's new InverseElementTransformation class in quest PointInCell
The InverseElementTransformation class allows setting different parameters for the
Newton-Raphson search, including, how to determine the initial isoparametric point
and what to do when the search leaves the element.

The PointInCell query now uses the closest physical point to the query point on a grid
of isoparametric points within the element and does not project the point back into
the element.

A future optimization might be to use the default options for InverseElementTransformation
on all candidate elements and to fine-tune the search (as above) if we cannot find the element
containing the point.