Added analytic cubic solver with tests.
The analytic cubic polynomial solver is based on Cardano's solution,
as set forth in various textbooks and papers (referenced in comments
in the source code).  This can get numerically unstable: the method
uses division by cube root of a number, which can rapidly propagate
noise.  The paper by Kopp cited in the source discusses that author's
work on eigenproblems of small matrices: use the analytic solution
with a fall-back to the slow, but very accurate, Jacobi's method.  We
may need to do something similar here.