ENH: add option to compile mint with sidre support
This commit adds a cmake option, MINT_USE_SIDRE,
that can be used to indicate whether to build
mint with Sidre support. This is a `cmake_dependent_option`
which defaults to ON when ENABLE_SIDRE is ON, otherwise,
it is set to OFF.

Note, using `cmake_dependent_option` required to
include the CMakeDependentOption module (provided by CMake)
in "src/CMakeLists.txt".

Last, the code that temporarily hard-coded MINT_USE_SIDRE is
removed. A config file, "mint/config.hpp", cofigured by CMake and
generated at build time, is included instead and consists of all
the appropriate definitions for mint.