ENH: add option to use 64-bit indices
This commit adds a mint cmake option, set to OFF by default, to
use 64-bit indices. A corresponding definition is added to
the mint/config.hpp that is configured and generated at build
time.

This commit also defines `mint::IndexType` which can be used
to index elements/nodes/fields on a mint::Mesh. Note,
`mint::IndexType` is defined in `mint/config.hpp` and is
aliased to a 32-bit integer or 64-bit integer according
to the corresponding cmake option.