BUGFIX: fixes for the  native Array constructors
This commit fixes minor issues with the array constructors. Notably,
the capacity of the Array must be at least `num_tuples*num_components`
in several places.

In addition, added another precondition check in the external buffer
constructor to ensure that the specified buffer is not NULL. The
signature of the external buffer constructor was also changed to
take in a `const T*` insted of a `T*`.

Last, improved the doxygen documentation for clarification and
documented pre- and post-conditions.