ENH: make num_tuples optional in Sidre constructor
This commit makes the "num_tuples" argument optional
when constructing an Array instance from an existing
sidre::View that has data.

The sidre::View has 2 dimensions, the first stores
the max capacity of the array, i.e., max number of
tuples,  and the second holds the number of componets
per tuple. When the max capacity is expected to be
equal with the actual number of tuples, the num_tuples
argument is not needed.

In addition, this commit improves the documentation for the
Sidre constructors and clarifies the pre- and post-conditions.

Last, minor stylist fixes are made in the doxygen style and
overall code.