ENH: refactor mint::MeshCoordinates
Refactors mint::MeshCoordinates to use mint::Array instead
of STL vector and integrates with Sidre according to the
computational mesh conduit blueprint protocol.

This commit enables the construction of a MeshCoordinates
object that may use (a) native storage for the coordinates, (b)
supplied external coordinates buffers or (c) from a sidre::Group
that conforms to the blueprint.

In addition, added comprehensive unit test for the
mint::MeshCoordinates class.

Changes to the rest of the classes to reflect this API changes
are also included.

This commit addresses in part ATK-945, ATK-958, ATK-942.