ENH: refactors mint::Field & mint::FieldVariable
Refactors the Field and FieldVariable classes to
use mint::Array internally instead of STL vector
and adds Sidre integration for mesh fields.

The API is revised to allow for a more streamlined
way of constructing and accessing mesh fields. A
field may now be constructed using (a) native storage
(b) an external, user-supplied buffer, or (c) a
Sidre::View that conforms to the mesh blueprint.

Changes throughout the Axom code reflecting the new
API changes are also included as part of this commit.

Last, comprehensive unit tests are added for the
mint::Field and mint::FieldVariable classes.

This commit addresses in part ATK-945, ATK-958, ATK-942.