ENH: generate a ccmake_cmd from config-build.py
Add the ability to generate a ccmake_cmd in the build-space
when the Axom CMake build is configured. This is for convenience
when using ccmake to interactively configure the build. We
typically use a spack installed cmake and a common error is
to try running ccmake on a build using a different version
of CMake (e.g., whatever is in the user's path). This
commit alleviates this problem.

Also, added checks to ensure that the specified `cmake` exists
and is executable.

If `ccmake` does not exist, i.e., it was disabled when building
cmake, a "ccmake_cmd" script is not generated in the build space.