Bugfix for spio_parallel test on Windows. ATK-1176.
The test was failing since all ranks were trying to read the file
pattern from the root file.

In the process, I created a utility function to broadcast a string from
rank 0 to all other ranks.

This bug was reminiscent of ATK-798 and was caught by the regression
test created for that bug (spio's write_read_write test).