BUGFIX: fix jacobian for surface elements
This commit fixes the way the jacobian is calculated
for a FiniteElement object. The initial implementation
was baking the assumption that the reference space of
the element had the same dimension as the ambient space
of the element. This does not hold true for surface elements,
wherein, the reference dimension is strictly lower than the
dimension of the ambient space, e.g., a quad or triangle in
3D space. Fixed that.

This commit partially addresses ATK-1052.