ENH: Add plane geometric primitive in Primal
This commit adds a plane geometric primitive in Primal.
The primal object defines an oriented plane in 2-D or 3-D
and provides support for plane operations, such as, projection
signed distance, orientation, etc. An associated unit
regression test is added as part of this commit.

In addition, this commit updates the CHANGELOG.md accordingly
to reflect this change.