ENH: refactors FieldData & Sidre integration
Adds the ability to create a FieldData object
from a Sidre fields group that conforms to the
mesh blueprint.

This commit also includes general improvements to
the FieldData API to streamline the creation of new
field variables and access to the field data in
application code. In addition, the ability to remove
field from a FieldData container is now implemented.

Comprehensive unit tests for the FieldData object
and changes throughout the Axom code reflecting the
new API changes are also included as part of this
commit.

This commit addresses in part ATK-945, ATK-958, ATK-942.