ENH: revert CellTypes enum class
Moving to a CellTypes enum class provided type safety,
but, complicated run-time querying for cell information
and required casting to index in to arrays

This commit reverts to a design consisting of a simple enum
that defines all the supported cell types and employs an
array of structs to hold information for each cell type.
Macros to register and query the cell information are
provided to keep the code cleand and simplify the process
of adding a new cell type.

Note, this commit temporarily comments out the ConnectivityArray
class and associated internals, since they are being refactored
and required additional changes for compilation.

This commit resolves in part ATK-958 and ATK-945