ENH: implements Mesh base class Sidre constructors
This commit implements the Mesh base class Sidre constructors
which are used by concrete instances. A new method is also
added which traverses the Sidre hierarchy and detects the
mesh type and dimension.

Note, currently the code will check/write some additional
metadata to store the BlockID and PartitionID of a mesh
object at the Blueprint root group.

This commit resolves in part ATK-945.