ENH: refactor ParticleMesh & sidre integration
This commit refactors the ParticleMesh/Mesh classes to
( 1 ) Remove use of STL
( 2 ) Revise and improve the API
( 3 ) Tighly integrate with Sidre

A ParticleMesh may now be constructed using native storage,
external buffers or from a Group in a Sidre hierarchy which
conforms to the computational mesh blueprint.

Comprehensive tests for the ParticleMesh class are also
included as part of this commit.

In addition, two examples are added the demonstrate how to
use the ParticleMesh class. A basic usage example, which
shows how to construct a ParticleMesh and simple N-Body
solver.

This commit also includes updates and a number of fixes to the Mesh
class that were encountered during development of the
ParticleMesh class, chief among them:

( 1 ) Handling for ParticleMesh in the generic access methods:
      Mesh::getMeshNode()
      Mesh::getMeshCell()
      Mesh::getCoordinateArray()

( 2 ) The code now emmits an error when the caller attempts to
      access anything other than node-centered data on a ParticleMesh

( 3 ) Small typos when navigating the Sidre hierarchy

Resolves in part ATK-945, ATK-958 and ATK-942.