ENH: Refactor UniformMesh & add sidre integration
This commit refactors the UniformMesh class to:
(a) remove use of STL from the internal implementation
(b) revise and improve the API
(c) tightly integrate with Sidre

A UniformMesh object may now be constructed using native storage
or with Sidre storage when Mint is compiled with Sidre support.
When using Sidre as the back-end datastore, Sidre owns all the
memory, in which case, the data remains persistent in Sidre
after the UniformMesh object is deallocated.

Support classes and methods, e.g., the StructuredMesh base class,
and the blueprint:: methods have been updated accordingly in support
of the functionality herein.

Comprehensive tests for the UniformMesh class are also included
as part of this commit. In addition, a  basic example that
illustrates how to use the UniformMesh class is added.

Last other code examples and operators have been updated to reflect
the API changes included by this commit.

Resolves in part ATK-945, ATK-958, ATK-942.