ENH: Refactor CurvilinearMesh & sidre integration
This commit refactors the CurvilinearMesh class to:
(a) remove use of STL form the internal implementation
(b) revise and improve the API
(c) tightly integrate with Sidre

A CurvilinearMesh object can now be constructed using native
storage, extenral storage, or Sidre, when Mint is compiled
with Sidre support. When using Sidre storage, Sidre owns all
the data, in which case, the data remains persistent in Sidre
when the CurvilinearMesh object is deallocated.

Support classes and methods are updated accordingly in support
of the functionality herein.

In addition, this commit adds comprehensive tests for the
CurvilinearMesh class and a simple example that illustrate
how to create and operate on a CurvilinearMesh.

Resolves in part ATK-945, ATK-958, ATK-942.