ENH: Refactor RectilinearMesh & sidre integration
This commit refactors the RectilinearMesh class. Specifically:
(a) removes use of STL from the internal implementation
(b) revises and improves the API
(c) tightly integrates with Sidre

A RectilinearMesh object may now be constructed using a native
storage constructor, an external storage constructor, or a
Sidre storage constructor, when Mint is compiled with Sidre
support.

Support classes and methods are updated accordingly in support
of the functionality herein.

In addition, this commit adds comprehensive tests for the
RectilinearMesh class and a simple basic usage example
that illustrates how to create and operate on a RectilinearMesh.

Resolves in part ATK-945, ATK-958, ATK-942.