ENH: decouple SignedDistance and In/Out Queries
This commit decouples the interface of the SignedDistance and In/Out
queries. These two queries, albeit similar, they are
different implementations, have different requirements and
take different parameters. Each query in Quest must provide
it's own exclusive C-style and/or C++ interface.

The new API also provides the means to programmatically set
different options via a scriptable API and can be extended
to accomodate new options as needed in the future.

Moreover, the init method now returns a status that indicates
whether the query has been initialized successfully to prevent
errors such as supplying a bad file name etc.

Last, tests and examples are added for the new interface.

This commit resolves ATK-1106, ATK-1099 and ATK-1098.