ENH: remove overloaded Matrix arithmetic operators
This commit removes the overloaded arithmetic operators from the Matrix class.
The overloaded arithmetic operators can have an undesirable performance
overhead. It is better if these operators are not provided to avoid
misuse in application code. Instead, application code may use the
corresponding free methods, which are less prone to misuse.

Moreover, this commits moves the free methods that implement various matrix
operators, e.g., matrix addition, subtraction, etc., into the matvecops
header which the designated location for these methods. These methods
are moved into the matvecops header and are renamed and refactored
accrodingly.

Tests and the rest of the Axom code are updated to reflect the API changes
introduced in this commit.