ENH: add return status to ops that can fail
Adds a return status to all matrix operators that can
fail due to undefined behavior, e.g., invalid matrix
dimension, etc. The return status is a boolean, which,
is set to true on success, otherwise, false.

All tests are  modified to reflect this change and
added new tests to check the status of operators when
not used properly.