BUGFIX: check for convergence on return
The jacobi_eigensolve() method was returning JACOBI_EIGENSOLVE_SUCCESS
even when the solver was not converging. This commit adds an appropriate
check for convergence before returning. The method will now return
JACOBI_EIGENSOLVE_FAILURE if the jacobi solver does not converge.
A corresponding unit test is also added as part of this commit.