ENH: support for MPI-3 shared memory in quest
This commit adds support for utilizing MPI-3, on-node shared memory
data-structures to store the input surface mesh to quest. Instead
of each rank duplicating the input mesh data, the ranks within the
same node can point to the same mesh data buffer, which is attached
to an MPI shared window. This reduces the memory overhead which can
be a limiting factor for very large meshes.

This commit resolves ATK-1074.