ENH: load STL to a Mesh with an external buffer
Adds support for reading in the STL mesh into a mint::Mesh
object that is bound to an external buffer. In this case,
the reader does not need to allocate memory for the mesh.
Instead it loads the data into the existing buffers. A
corresponding test is added to the STL reader unit test.