ENH: add option to toggle use of MPI 3 features
Adds a cmake option to enable or disable the use of MPI-3
features, e.g., shared memory, etc. This option is dependent
on whether ENABLE_MPI is set to true, it defaults to OFF and is
marked as advanced.

A corresponding cmakedefine is added to the config.hpp.in
which gets configured accordingly by CMake.