Changes for Shroud-0.10.0
Order of generated classes changed due to how forward declarations are
handled.

The C API now generates a struct with the C++ class pointer and some
information about how to deallocate the pointer for shadow classes.
C++ wrappers which return a class instance now require an additional
argument for the pointer to the struct, which is also returned
(similar to how strcpy returns a pointer to the dest argument) to
allow it to be used in expressions.  The user must pass in the the
pointer to the struct to avoid complications that would be introduced
if the struct was allocated by the C wrapper.  The Fortran derived
type for the class contains the struct which is passed down to the C
wrapper.  The header files type<LIBRARY>.h are created which contain
the struct definition.

Syntax for template instantiation changed.

Add ability to update typemap directly from YAML file.  Used to update
MPI_Comm for quest.