ENH: simplify external storage constructors
Simplifies the unstructured mesh external storage constructors.
The caller now does not need to specify the mesh dimension since
that can be internally detected from the supplied coordinate
buffers.

Also, added delegating constructors for the external storage
constructors, such that, the node_capacity and cell_capacity
of the supplied buffers does not have to be explicitly specified
since in the common case the capacity of the supplied buffers
is the same with the specified size.