BUGFIX: disable death tests when mpi is ON
The signed_distance_interface test can be compiled both
with and without MPI. When MPI is turned on, SLIC_ERROR()
calls MPI_Abort(), which terminates all processes. The behavior
of MPI_Abort() varies depending on the implementation. So, let's
just not do that when MPI is enabled for this test.