Adds note to 'update_copyright_date.sh' about missed sphinx config files
These need to be hand-edited for now, or folded into the script.

They currently do not satisfy either the filename regex
(since they do not have the line  'This file is part of Axom')
or the sed regex (since they don't have the words 'Copyright (c)'
on the line to be replaced.