Fixes a bug in quest's vertex welding application for small epsilon values (ATK-1291)
We were using 32 bit integers for the indices, but this only allows 2^31 ~ 2 billion grid cells
per dimension in the RectangularLattice. However, for small values of epsilon (e.g. EPS = 1E-13),
this leads to a bounding box range of 0.0002 units.

In contrast, 64 bit integers support a range of about a million units (i.e 2^63 * EPS ).