Added a script to make a gzipped release tarball.
When you make a release on Github, the tarfiles it generates don't properly
include Git submodules. This script generates a gzipped tarfile named for
the release that can be linked to the Github release for users to download.

It strips out all the git stuff using 'git archive'.

Here's the recipe to use the script:

1) Edit the script and set the 'VERSION' variable to the release number. This
   is an extra step, but it's easy and makes the script useful for other things
   such as generating a complete tarfile of Axom from any commit and give it
   to someone.
2) In the top-level Axom directory, type: ./scripts/make_release_tarball.sh
3) The generated gzipped tarfile in the top-level Axom directory is good to go.
   Use it like you would a clone of the RAJA repo.