ENH: adds mark boundaries in watertight check
Adds an optional flag to the "isSurfaceMeshWaterTight"
method to mark the cells that are on a boundary. When the caller
specifies to mark boundary cells, a cell-centered
field variable is added, called "boundary", whose
value is "1" for cells on the boundary, "0" otherwise.
This allows to easily visualize the celss found to be on
the boundary and detect the regions that cause problems
to the In/Out and SignedDistance queries.