BUGFIX: treat core_types test as an MPI test
This commit changes the way the core_types test is compiled
and added to the test suite. When MPI is disabled, the test
is added with all the other tests. In contrast, when MPI is
enabled, the core_types test is linked with MPI and added
to the test suite as an MPI test that runs on a single rank.

This would ensure that the test is properly srun'ed on LC
platforms.