COMP: fix for mpi_traits on blueos
Resolves build issues on blueos with the mpi_traits struct.
Depending on the MPI implementation, MPI_Datatype may not
be a type that can be defined as a static constexpr member
of a class or struct and initialized inline. This was leading
to build failures on blueos, specifically with the coral XL and
clang compilers.

This commit fixes the issue by defining MPI_Datatype as
a static const member of the mpi_traits struct and initializing
the defaults for each specialization in a seperate CPP file.