ENH: expand mint getting started example
This commit expands Mint's getting started example as follows:

( 1 ) Adds support for unstructured meshes as well. The mesh type
can be selected at runtime via a command line switch. This
illustrates how Mint can be leveraged to develop kernels that
are both mesh-agnostic and device-agnostic.

( 2 ) Illustrates how the "nodeIds" can be used in the second kernel
to average a node-centered field in a corresponding cell-centered
quantity.

( 3 ) Adds the ability to set the mesh resolution at the command line.