COMP: fix BG/Q cmake build errors
The BG/Q build was failing when CMake was trying to configure
due to using undefined variables in conditional logic.

Specifically, checks of the form:
```
if ( ${RAJA_FOUND} AND ${UMPIRE_FOUND} )
    ...
endif()
```

Would lead to cmake configuration error when RAJA_FOUND and
or UMPIRE_FOUND are undefined.

Change these checks to:
```
if ( RAJA_FOUND AND UMPIRE_FOUND )
    ...
endif()
```

which works properly when the variables are undefined as well.

This commit fixes the issue encountered on BG/Q where we don't
specify an UMPIRE_DIR.