ENH: exclude Axom/data from release archive
This commit changes the release tarball script such that
the Axom/data submodule is not included with the source
release tarball. Instead, the "--with-data" option may
be supplied to the script, which would create a different
tarball consisting of the corresponding snapshot of the
AxomData that is in use.

This commit resolves #74.