BUGFIX: usage of umpire's MemoryResourceType enum
Axom was implicitly making the assumption that umpire's
MemoryResourceType enum, e.g., "umpire::resource::Host",
"umpire::resource::Unified", etc., corresponded to the
allocator ID of the predefined allocators for the associated
resource. This used to be the case in earlier versions of umpire,
but, generally it cannot be guaranteed, leading to errant
behavior, in CUDA-enabled builds in particular.

This commit adds the ability to query the allocator ID of
the predefined allocator, created by umpire, for a particular
resource and corrects the usage of the MemoryResourceType enum
throughout Axom.