ENH: generic BVH traversal routine
This commit adds a generic routine for traversing the BVH where the
inLeft, inRight predicates, as well as, the leaf node action is
prescribed by the caller as a lambda or functor. This enables to
minimize the code duplication in the BVH::find() routine and
streamlines the extension of the BVH to support additional queries.

This commit resolves in part #135.