BUGFIX: fix failing mint test with intel18-opt
The `mint_fem_single_fe` test was faling in Release builds of
Axom (i.e, optimized builds) using the intel18 compiler. The
compiler optimizations were leading to erroneous results in
the Matrix::swapColumns() due to a compiler bug. Notably, this
behavior is not reproducible with intel19. Inlining the "swap"
operation instead of calling utilities::swap()
circumvents this issue.

This commit resolve issue #90.