Bugfix for truncation feature in 'convert_sidre_protocol' utility
The original algorithm did not properly handle cases where there were multiple
views pointing to the same data. To fix this, we now create a copy of the
truncated data.

We also prepend to each array the number of retained samples in addition to the
size of the original array.