ENH: use CudaKernelFixed and tread_direct policies
Raja changed the way cuda kernel launch bounds are calculated, which
was leading to a large number of threads being launched in the way
policies were defined previously. The large number of threads was
bloating the number of registers being used leading to a linking error
due max reg count violation. Fixed that by using Raja CudaKernelFixed
in conjuction with the "thread direct" policies.