ENH: define TraversalPredicates singleton
This commit adds a singleton class consisting of methods that implement
various traversal predicates, used in support of various queries. The singleton
works directly on the internal BVHData layout to process BVH bins and
is specialized for 2D and 3D.

Currently, it supports point-in-bin queries. However, this provides the
underlying infrastructure to easily extend the BVH for support of
additional queries.

This commit resolves in part #135.