ENH: add portable floating point traits class
Add standardized and portable floating point traits class
that allows querying lowest(), min() and max() values for
a given floating point on either host or device. The functionality
that this class provides is equivalent to that of std::numeric_limits,
however, these functions are host/device decorated and therefore
can be called from device code as well.