BUGFIX: properly clear shared buffer in finalize
When enabling MPI-3 shared memory for the signed distance query,
signed_distance_finalize() deallocates the on-node shared buffer
by MPI_Win_free(). However, the value of the pointer used to
access that buffer internally in the code was not being properly
set to null. Consequently, subsequent calls to quest::initialize()
would fail.

This resolves #220.