Changes for Shroud-0.12
A dimension attributes must now have a shape. Instead use +rank(1) for
1-d input arrays.

Functions which initialize shadow classes are now a generic interface
with the derived type name.
  datastore_new -> SidreDataStore
  iomanager_new -> IOManager
Fortran is case insensitive, but use the same case as C++ for tests.

Change prototypes for function which accept a shape array to be 'const
IndexType *'.  Shroud defaults this to intent(in).