BUGFIX: workaround for clang@9.0.0 segfault
The clang@9.0.0 compiler was segfaulting when compiling
the mint_cell_types.cpp unit test. The key issue that
was triggering this is the use of `constexpr` when
defining the static cell_info array in CellTypes.hpp.
This appears to be a bug in the clang@9.0.0 compiler.

Althought it would be nice to use `constexpr` for the
cell info, Mint does not heaviliy rely on this. To
workaround this issue, the code has been changed to
use `const` instead of `constexpr`.

In addition, the `mint_cell_types.cpp` unit test is
now re-enabled for all specs.

This commit resolves #266