BUGFIX: fix for L1 incoherent cache on NVIDIA GPUs
This commit attempts to fix an issue with the parallel
construction of the BVH on GPUs that is resulting due
to the L1 cache not being coherent across all GPU threads.
Specifically, it adds a call to the ``__threadfence_system()``
in the ``propagate_aabbs`` method after a parent BVH node is
computed and stored in global memory to ensure that the "write"
will be visible by all threads in the system.

This commit resolves #307