BUGFIX: don't call umpire outside of main
Due to the multi-gpu support that was recently added
in Umpire, the umpire::ResourceManager now makes
calls to CUDA --  specifically, to ensure that memory
is accessible from all visible devices. CUDA calls
must happen after CUDA is properly "initialized".
Consequently, the umpire::ResourceManager must be
instantiated after CUDA is "initialized". In the present
implementation we are initializing a global const variable,
DEFAULT_ALLOCATOR_ID, at the top of a header, by instantiating
the `umpire::ResourceManager`. This unfortunately, triggers
CUDA calls, before CUDA is "initialized", which leads to
strange errors in all of our CUDA tests: launching CUDA kernels
fails.

This commit provides a work-around to this issue by avoiding
calling the umpire::ResourceManager outside of main.