Use CMake's generate_export_headers to deal with importing/exporting symbols in dlls
Note: CMake is supposed to add a `DEFINE_SYMBOL` to each target of the form `<target>_EXPORTS`,
but this symbol was not defined in MSVC, leading to import/export linker warnings and errors.
I suspect that the problem might be due to our usage of object libraries.
I fixed this by explicitly adding `axom_EXPORTS` as a private compiler define for each axom component.