Workaround for C++14 feature missing from gcc@4.9.3
Some user configurations are built with C++14, but are pointing at a gcc@4.9.3 standard library
The latter is not fully C++14 compliant, but is the default on some platforms,
and it can be difficult to propagate the right flags in some build configuration.
This change allows the code to build in the above configuration.