Adds a struct to accelerate finding the starting element for a point insertion query
Uses a naive regular grid of bins to cache to index of a vertex.
The heuristic is to use `2 * N^1/(2D)` bins in each dimension when there are N points.
E.g. for 1,000,000 vertices, this gives around 60x60 bins in 2D, each with ~125 vertices
and 20x20x20 bins in 3D, each with ~250 vertices.

Without this, we were using an arbitrary element (e.g. the most recently inserted)
and insertions into large meshes were significantly slower.
With this, the insertion rate is reasonably maintained for meshes with 1,000,000 and 10,000,000
meshes.

We will likely see better acceleration with an Octree instead of a regular grid.
This would come at the cost of more memory for the spatial index (Octree vs. regular grid)
and slightly more expensive lookups.