Use a UniformGrid for vertex queries in Delaunay::isValid()
The previous implementation was using an ImplicitGrid.
For each of vertex in a mesh with V vertices and E elements,
it was getting a Bitset of O(E) bits, even though there are,
on average, only O(1) elements that we actually need to check.
When V and E are large, e.g. when V is O(E^6), it was taking 10-20x
longer to validate the Delaunay complex than to generate it incrementally.

The current approach is to use an ImplicitGrid to bootstrap the index
and then copy over the precomputing intersections to a UniformGrid.
This avoids excessive allocations associated with insertions into a UniformGrid.