Bugfix: In `quest::SignedDistance`, need to unitize edge normals
Without unitizing, we can run into problems for triangles with different sizes (areas)
when the dihedral angle between the triangles is acute (less than 90 degrees).
Specifically, without this fix, the larger triangles will add a larger weight to the normal
and some regions closer to the smaller triangle can get the wrong sign
in `SignedDistance::computeSigns()`.