Set axom::Sidre::IndexType to axom::IndexType
Add a typedef for TypeID in C as SIDRE_TypeID and in Fortran as
TypeID.  By not having a leading SIDRE_ in the Fortran variable it
works around a bug in Shroud that does not add continuations on the
IMPORT statement.

SIDRE_TypeID is set to short in C and Fortran to allow some overloaded
functions to be distinguished. In C++ the enum is a different type
than IndexType. But in Fortran, if they are all C_INT then the
compiler has issues selecting the specific function for the generic
function view::apply.

subroutine view_apply_nelems_offset_stride(obj, num_elems, offset, stride)
subroutine view_apply_type_nelems_offset(obj, type, num_elems, offset)

A short variable is large enought to hold all of the TypeIDs that we
have.