Work around amdflang error with sizeof TypeID
The Fortran wrappers converts the enum TypeID into a short.  This
allows some generic function to be able to be distinguished when
IndexType is C_INT32_T. This makes TypeID a different kind than
IndexType.

However, amdflag on corona reports an error:

    wrapfsidre-d0d02c.ll:138:304: error: invalid forward reference to
    function 'axom_sidre_buffer_get_type_id_' with wrong type: expected
    'i32 (i64*, i64*)*' but was 'i16 (i64*, i64*)*'

The methods getTypeID return an axom::sidre::TypeID.  Added a typemap
to sidre_shroud.yaml similar to the TypeID typemap, named TypeIDint,
which will use C_INT in the Fortran (and C) wrapper for the return
value of getTypeID.  Arguments which accept a TypeID argument continue
to use a C_SHORT.

Once the compiler is fixed, TypeIDint can be returned to TypeID.